pyproject.toml
[tool.poetry]
name = "pyinstaller_versionfile"
version = "3.0.0"
description = "Create a windows version-file from metadata stored in a simple self-written YAML file or obtained from an installed distribution."
authors = ["Andreas Finkler"]
license = "MIT"
readme = "README.md"
include = ["CHANGELOG.md", "LICENSE", "tox.ini", "pylintrc", "test"]
repository = "https://github.com/DudeNr33/pyinstaller-versionfile"
homepage = "https://github.com/DudeNr33/pyinstaller-versionfile"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Topic :: Software Development",
    "Topic :: Software Development :: Build Tools",
    "Environment :: Console",
]

[tool.poetry.scripts]
create-version-file = "pyinstaller_versionfile.__main__:create_version_file"
pyivf-make_version = "pyinstaller_versionfile.__main__:make_version"

[tool.poetry.dependencies]
python = "^3.9"
Jinja2 = "*"
PyYAML = "*"
packaging = "*"

[tool.poetry.group.dev.dependencies]
setuptools = "^74.1.2"
pylint = "^3.2.7"
pytest = "^8.3.3"
tox = "^4.18.1"
black = "^24.8.0"
pyinstaller = { version = "^6.10.0", python = "<3.14" }             # PyInstaller unfortunately does not use "<4.0" as upper bound - this Python restriction will need to be updated once in a while
pywin32 = { version = "^306", markers = "sys_platform == 'win32'" }
mypy = "^1.15.0"
types-pyyaml = "^6.0.12.20250402"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
